Fix benchmarks (& examples) where sizes exceed max int32 (#313)
The `--m=8` benchmarks fail because when we do this:
```cpp
block_B.reset(N * K * L);
```
the product exceeds max int32, so we allocate the wrong amount of
memory.

We also need to disable the sycl index check:
`-fno-sycl-id-queries-fit-in-int`

In this PR I have:
 - set that flag in FindDPCPP.cmake
- ensured that all allocation sizes (benchmarks & examples) are in
`std::size_t`
 - re-enabled the benchmarks which these changes fix

---------

+++ examples/sycl/02_pvc_gemm_mixed_dtype/02_pvc_gemm_mixed_dtype.cpp
@@ -391,15 +391,15 @@ struct ExampleRunner {
     stride_D = cutlass::make_cute_packed_stride(StrideD{}, shape_CD);
     stride_S = cutlass::make_cute_packed_stride(StrideScale{}, shape_scale_zero);
 
-    block_A.reset(M * K * L);
-    block_A_dq.reset(M * K * L);
-    block_B.reset(K * N * L);
-    block_B_dq.reset(K * N * L);
-    block_C.reset(M * N * L);
-    block_D.reset(M * N * L);
-    block_ref_D.reset(M * N * L);
-    block_scale.reset(scale_k * L * dq_mn_size);
-    block_zero.reset(scale_k * L * dq_mn_size);
+    block_A.reset(static_cast<std::size_t>(M) * K * L);
+    block_A_dq.reset(static_cast<std::size_t>(M) * K * L);
+    block_B.reset(static_cast<std::size_t>(K) * N * L);
+    block_B_dq.reset(static_cast<std::size_t>(K) * N * L);
+    block_C.reset(static_cast<std::size_t>(M) * N * L);
+    block_D.reset(static_cast<std::size_t>(M) * N * L);
+    block_ref_D.reset(static_cast<std::size_t>(M) * N * L);
+    block_scale.reset(static_cast<std::size_t>(scale_k) * L * dq_mn_size);
+    block_zero.reset(static_cast<std::size_t>(scale_k) * L * dq_mn_size);
 
     initialize_mixed_dtype_block(block_A, block_A_dq, seed + 2023);
     initialize_mixed_dtype_block(block_B, block_B_dq, seed + 2022);
